+++ services/core/methods/controllers/accounts.js
@@ -43,12 +43,9 @@ const getKnownAccounts = async () => {
 };
 
 const getDataForAccounts = async params => {
-	let accounts;
-	if (params.isDelegate === 'true') {
-		accounts = await CoreService.getDelegates(params);
-	} else {
-		accounts = await CoreService.getAccounts(params);
-	}
+	const accounts = params.isDelegate
+		? await CoreService.getDelegates(params)
+		: await CoreService.getAccounts(params);
 
 	const response = {};
 	response.data = [];
