+++ services/core/shared/core/compat/sdk_v5/coreCache.js
@@ -13,9 +13,7 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const { Utils, CacheLRU } = require('lisk-service-framework');
-
-const { isEmptyArray } = Utils.Data;
+const { CacheLRU } = require('lisk-service-framework');
 
 const config = require('../../../../config');
 const { knex } = require('../../../database');
@@ -29,12 +27,12 @@ const getCachedAccountBy = async (key, value) => {
 	let account = await cache.get(cacheKey);
 	if (!account) {
 		const result = await accountsDB.find({ [key]: value });
-		if (!Array.isArray(result) || isEmptyArray(result)) {
+		if (result === undefined) {
 			const expireMiliseconds = config.ttl.affectedByNewBlocks;
 			await cache.set(cacheKey, null, expireMiliseconds);
 			return null;
 		}
-		const { address, username, publicKey } = result[0];
+		const [{ address, username, publicKey }] = result;
 		account = { address, username, publicKey };
 		Object.entries(account).forEach(async ([k, v]) => {
 			if (v) await cache.set(getCacheKey(k, v), account);
